Update `bin/replace` script for better portability
The existing implementation of the `replace` script does not work on
GNU/Linux due to differences in the implementation of `sed`. This change
reworks the process of the replacement to create temporary files,
apply the changes to the files, and then move them into place. It will
improve the script's portability for more OSes.

--
Overwrite internal field separator when replacing.
Problem:
Currently the `for` in bash will break up lists by whitespace, so if a
filename has a space in it the `for` will break it up incorrectly.

AG will separate the files with a newline character, so this temporarily
overwrites the `IFS` to be used correctly for this use case.

[Docs](https://bash.cyberciti.biz/guide/$IFS)

* Commits imported from Thoughtbot:
- https://github.com/thoughtbot/dotfiles/commit/2d9e8bf0e84a835f300bd519faef4b6bfe289a47
- https://github.com/thoughtbot/dotfiles/commit/1dc550d9850337594401349ea0ceeb922beef405