Add vim plugin tmux-plugins/vim-tmux-focus-events
`FocusGained` and `FocusLost` autocommand events are not working in
terminal vim. This plugin restores them when using vim inside Tmux.

Here's where that matters:

- tpope/vim-fugitive plugin uses `FocusGained` for refreshing git branch
  in status line
- airblade/vim-gitgutter uses `FocusGained` for refreshing the git gutter

Also, `vim-tmux-focus-events` makes the `autoread` option work properly
for terminal vim. So far, this was only working in a GUI version.

The `autoread` feature comes handy when files are changed outside vim, for
example when resolving merge conflicts. When you come back to vim and try saving
a changed file you'll likely be interrupted with E813*.

Improved `autoread` prevents this by automatically reading a file from disk
if it was changed. Works only if `autoread` option is set (enable it with
`set autoread` in `.vimrc`).

* http://vimdoc.sourceforge.net/htmldoc/editing.html#E813